reflection: Split type_traits from adl
The type traits are useful without the overhead of iobuf.
